requirements.txt
fastapi
uvicorn
telethon
beautifulsoup4
requests
python-dotenv
pydantic
httpx
